+++ src/components/Header.js
@@ -1,6 +1,13 @@
+import { Link, NavLink } from "react-router-dom";
 import logo from "../assets/images/logo/sm-logo.svg";
-import typoLogo from "../assets/images/logo/type-logo.svg";
+import typoLogo from "../assets/images/logo/typo-logo.svg";
 import "./Header.css";
+
+const getLinkStyle = ({ isActive }) => {
+  return {
+    color: isActive ? "#3692FF" : undefined,
+  };
+};
 
 function Header() {
   return (
@@ -8,13 +15,17 @@ function Header() {
       <div>
         <picture>
           <source srcSet={typoLogo} media="(max-width: 743px)" />
-          <a href="index.html">
+          <Link to="/">
             <img src={logo} alt="판다마켓 로고" />
-          </a>
+          </Link>
         </picture>
         <div className="headerMenu">
-          <p>자유게시판</p>
-          <p>중고마켓</p>
+          <NavLink to="/board" style={getLinkStyle} className="link">
+            <p>자유게시판</p>
+          </NavLink>
+          <NavLink to="/items" style={getLinkStyle} className="link">
+            <p>중고마켓</p>
+          </NavLink>
         </div>
       </div>
       <button className="profileBtn" type="button">
